Adopt commit 25e8d7a from master: Merge pull request #326 from jmendelsohn4/bdl_build_warnings-drqs-81106581
commit 25e8d7aef5a9e56e87faabe454c5fbeaad073e6d
Merge: 599f226 1dfb731
Author: Jeffrey Mendelsohn <JMENDELSOHN4@bloomberg.net>
Date:   Mon Apr 18 10:12:12 2016 -0400

    Merge pull request #326 from jmendelsohn4/bdl_build_warnings-drqs-81106581

    fix bdl compile warnings blocking nightly build

commit 1dfb73154aae31bc9ce297ed81f3d83831b39844
Author: Jeffrey Mendelsohn <JMENDELSOHN4@bloomberg.net>
Date:   Fri Apr 15 14:42:39 2016 -0400

    incorporating HV feedback

commit 4ac07aa234607cb43271e180f0747da7658238eb
Author: Jeffrey Mendelsohn <JMENDELSOHN4@bloomberg.net>
Date:   Fri Apr 15 13:38:36 2016 -0400

    incorporated AM review comment

commit 3026c2727e325e8ab4fd0131b417df55d1886698
Author: Jeffrey Mendelsohn <JMENDELSOHN4@bloomberg.net>
Date:   Fri Apr 15 13:00:58 2016 -0400

    fixing compiler warnings; this also lead to a change in skiplist; I am aware of the good-pull-request policy violation

commit 8de4b730295f5e187244d833c3df27c46d04c1c7
Author: Jeffrey Mendelsohn <JMENDELSOHN4@bloomberg.net>
Date:   Fri Apr 15 08:25:27 2016 -0400

    fix bdl compile warnings blocking nightly build

+++ groups/bdl/bdlat/bdlat_typename.cpp
@@ -237,7 +237,7 @@ bool bdlat_TypeName_Imp::idempotentConcat(char       *dest,
     int remaining = destSize - 1;
     for (int i = 0; i < numSegments; ++i) {
         const char *segment = segments[i];
-        int segmentLen = bsl::strlen(segment);
+        int segmentLen = static_cast<int>(bsl::strlen(segment));
         if (segmentLen > remaining) {
             segmentLen = remaining;
         }
