+++ groups/bdl/bdlt/bdlt_calendar.h
@@ -1691,7 +1691,8 @@ int Calendar::getNextBusinessDay(Date *nextBusinessDay, const Date& date) const
 
     enum { e_SUCCESS = 0, e_FAILURE = 1 };
 
-    int offset = d_nonBusinessDays.find0AtMinIndex(date + 1 - firstDate());
+    int offset = static_cast<int>(
+                    d_nonBusinessDays.find0AtMinIndex(date + 1 - firstDate()));
     if (0 <= offset) {
         *nextBusinessDay = firstDate() + offset;
         return e_SUCCESS;                                             // RETURN
@@ -1766,7 +1767,7 @@ int Calendar::length() const
 inline
 int Calendar::numBusinessDays() const
 {
-    return d_nonBusinessDays.num0();
+    return static_cast<int>(d_nonBusinessDays.num0());
 }
 
 inline
@@ -1776,8 +1777,8 @@ int Calendar::numBusinessDays(const Date& beginDate, const Date& endDate) const
     BSLS_ASSERT_SAFE(isInRange(endDate));
     BSLS_ASSERT_SAFE(beginDate <= endDate);
 
-    return d_nonBusinessDays.num0(beginDate - firstDate(),
-                                  endDate - firstDate() + 1);
+    return static_cast<int>(d_nonBusinessDays.num0(beginDate - firstDate(),
+                                                   endDate - firstDate() + 1));
 }
 
 inline
@@ -1801,7 +1802,7 @@ int Calendar::numHolidays() const
 inline
 int Calendar::numNonBusinessDays() const
 {
-    return d_nonBusinessDays.num1();
+    return static_cast<int>(d_nonBusinessDays.num1());
 }
 
 inline
@@ -2019,8 +2020,8 @@ Calendar_BusinessDayConstIter& Calendar_BusinessDayConstIter::operator++()
 {
     BSLS_ASSERT_SAFE(d_currentOffset >= 0);
 
-    d_currentOffset =
-                     d_nonBusinessDays_p->find0AtMinIndex(d_currentOffset + 1);
+    d_currentOffset = static_cast<int>(
+                    d_nonBusinessDays_p->find0AtMinIndex(d_currentOffset + 1));
     return *this;
 }
 
@@ -2028,12 +2029,12 @@ inline
 Calendar_BusinessDayConstIter& Calendar_BusinessDayConstIter::operator--()
 {
     if (-1 == d_currentOffset) {
-        d_currentOffset = d_nonBusinessDays_p->
-                             find0AtMaxIndex(0, d_nonBusinessDays_p->length());
+        d_currentOffset = static_cast<int>(d_nonBusinessDays_p->
+                            find0AtMaxIndex(0, d_nonBusinessDays_p->length()));
     }
     else {
-        d_currentOffset = d_nonBusinessDays_p->
-                                           find0AtMaxIndex(0, d_currentOffset);
+        d_currentOffset = static_cast<int>(d_nonBusinessDays_p->
+                                          find0AtMaxIndex(0, d_currentOffset));
     }
 
     BSLS_ASSERT_SAFE(0 <= d_currentOffset);
@@ -2110,7 +2111,7 @@ struct UsesBslmaAllocator<bdlt::Calendar> : bsl::true_type {};
 #endif
 
 // ----------------------------------------------------------------------------
-// Copyright 2015 Bloomberg Finance L.P.
+// Copyright 2016 Bloomberg Finance L.P.
 //
 // Licensed under the Apache License, Version 2.0 (the "License");
 // you may not use this file except in compliance with the License.
